QUAL-194: Updated build script with 'norpm' parameter and changed parameter implementation

+++ build/buildqual.sh
@@ -1,24 +1,44 @@
 #!/bin/bash
 
+BUILD="QUAL"
+RPM="YES"
+
 # Check for parameters, 
-# if none, build only qual image 
-# if vm,   build only vm image 
-# if all,  build both
-if   [ $# -eq 0 ];  then BUILD="QUAL"
-elif [ $1 == vm ];  then BUILD="VM"
-elif [ $1 == all ]; then BUILD="ALL"
-else
-    echo "Unrecognized parameter specified.  Accepted parameters are:
-    all - builds both qual and qual-vm images
-     vm - builds only qual-vm image"
-    exit
-fi
+# if none, 	build only qual image 
+# if norpm, 	build images without re-building RPMs
+# if vm,   	build only vm image 
+# if all,  	build both
+TEMP=`getopt -o nva --long norpm,vm,all -n 'buildqual.sh' -- "$@"`
+eval set -- "$TEMP"
+
+while true ; do
+    case "$1" in
+	-n|--norpm)
+	    RPM="NO"
+	    shift;;
+	-v|--vm)
+	    BUILD="VM"
+	    shift;;
+	-a|--all)
+	    BUILD="ALL"
+	    shift;;
+	--)
+	    shift; break;;
+	*)
+    	    echo "Unrecognized parameter specified.  Accepted parameters are:
+    		-n|--norpm 	- builds images without re-building RPMs
+     		-v|--vm 	- builds only qual-vm image
+		-a|--all 	- builds both qual and qual-vm images"
+    	    exit 1;;
+    esac
+done
 
 QUALSRCDIR=/home/thales/qual/src
 MPSBUILDDIR=/home/thales/mps-builder
 
 # Handle tito tag and build for qual
 titoqual () {
+    echo "Building qual RPM! ('-' )"
     cd ${QUALSRCDIR}/
     tito tag
     tito build --rpm --offline
@@ -26,6 +46,7 @@ titoqual () {
 
 # Handle tito tag and build for qual-vm
 titovm () {
+    echo "Building qual-vm RPM! (._. )"
     cd ${QUALSRCDIR}/simulator/
     tito tag
     tito build --rpm --offline
@@ -33,6 +54,7 @@ titovm () {
 
 # Build qual pxe image
 buildqual () {
+    echo "(/*-*)/ Building qual images! \(*-*\)
     sudo cp ${QUALSRCDIR}/../build/pkgs-qual.inc.ks ${MPSBUILDDIR}/config/
     sudo docker run --net=host --rm=true -u root --privileged=true -v ${MPSBUILDDIR}:/mnt/workspace -v /dev:/dev -t mps/mpsbuilder:centos7 /bin/bash "/mnt/workspace/build.script"
     cd ${MPSBUILDDIR}/bin/
@@ -43,6 +65,7 @@ buildqual () {
 
 # Build qual-vm pxe image
 buildvm () {
+    echo "(/~-~)/ Building qual-vm images! \(~-~\)
     sudo cp ${QUALSRCDIR}/../build/pkgs-qual-vm.inc.ks ${MPSBUILDDIR}/config/pkgs-qual.inc.ks
     sudo docker run --net=host --rm=true -u root --privileged=true -v ${MPSBUILDDIR}:/mnt/workspace -v /dev:/dev -t mps/mpsbuilder:centos7 /bin/bash "/mnt/workspace/build.script"
     cd ${MPSBUILDDIR}/bin/
@@ -58,16 +81,21 @@ buildvm () {
 }
 
 set -e
-cd ${QUALSRCDIR}/
-echo "Please use your own Git credentials to log in. \(^^\) \(^^)/ (/^^)/"
-git fetch origin dev/QUAL
-git reset --hard FETCH_HEAD
-git clean -df
-rm -rf /tmp/tito
-tito init
-titoqual
 
-if [ $BUILD != "QUAL" ]; then titovm; fi
+if [ $RPM == "YES" ]; then
+    cd ${QUALSRCDIR}/
+    echo "Please use your own Git credentials to log in. \(^^\) \(^^)/ (/^^)/"
+    git fetch origin dev/QUAL
+    git reset --hard FETCH_HEAD
+    git clean -df
+    rm -rf /tmp/tito
+    tito init
+    titoqual
+
+    if [ $BUILD != "QUAL" ]; then titovm; fi
+
+    git push origin dev/QUAL
+fi
 
 sudo rm -f ${MPSBUILDDIR}/repo/packages/x86_64/qual-*.rpm
 sudo mv /tmp/tito/x86_64/* ${MPSBUILDDIR}/repo/packages/x86_64/
@@ -78,9 +106,6 @@ case $BUILD in
       "VM") buildvm;;
      "ALL") buildqual; buildvm;;
 esac
-
-cd ${QUALSRCDIR}/
-git push origin dev/QUAL
 
 if [ $BUILD != "VM" ]; then 
     echo "Built QUAL PXE image: $QUALPXE"
